fixed smell code + renamed files
