+++ src/distance_forest.py
@@ -1,23 +1,28 @@
+"""
+Create a json   
+"""
+
 from math import radians, sin, cos, sqrt, atan2
 from json import dumps, loads
 
 
-def distance(lat1, lon1, lat2, lon2) -> float:
+def distance(latitude1, longitude1, latitude2, longitude2) -> float:
     """ Calculates the distance between 2 coordinates"""
     # radius of the Earth
-    R = 6371
+    earth_radius = 6371
 
     # degree to radians conversion
-    lat1, lon1, lat2, lon2 = map(radians, [lat1, lon1, lat2, lon2])
+    latitude1, longitude1, latitude2, longitude2 = map(radians,
+                                                [latitude1, longitude1, latitude2, longitude2])
 
     # latitude and longitude differences
-    dlat = lat2 - lat1
-    dlon = lon2 - lon1
+    dlat = latitude2 - latitude1
+    dlon = longitude2 - longitude1
 
     # Haversine formula
-    a = sin(dlat/2)**2 + cos(lat1) * cos(lat2) * sin(dlon/2)**2
-    c = 2 * atan2(sqrt(a), sqrt(1-a))
-    dist = R * c
+    a_value = sin(dlat/2)**2 + cos(latitude1) * cos(latitude2) * sin(dlon/2)**2
+    c_value = 2 * atan2(sqrt(a_value), sqrt(1-a_value))
+    dist = earth_radius * c_value
 
     return dist
 
@@ -34,21 +39,21 @@ for forest in forests:
     lat1 = forest["coords"][1]
     long1 = forest["coords"][0]
 
-    closest_city = None
-    distance_ = None
+    CLOSEST_CITY = None
+    DISTANCE = None
 
     for city in cities:
         lat2 = float(city["latitude"])
         long2 = float(city["longitude"])
         d = distance(lat1, long1, lat2, long2)
-        if closest_city is None or distance_ > d:
-            closest_city = city["label"]
-            distance_ = d
+        if CLOSEST_CITY is None or DISTANCE > d:
+            CLOSEST_CITY = city["label"]
+            DISTANCE = d
 
     cities_forests.append({
-        "city" : closest_city, 
+        "city" : CLOSEST_CITY, 
         "forest" : forest["nom"],
-        "distance" : distance_
+        "distance" : DISTANCE
     })
 
 if __name__ == "__main__":
